@@ -2,10 +2,10 @@ package com.hendisantika.service;
 
 import com.hendisantika.entity.Employee;
 import com.hendisantika.repository.EmployeeRepository;
+import jakarta.annotation.PostConstruct;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import javax.annotation.PostConstruct;
 import java.util.List;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
